Properly ignore contrain errors on make_active
In general this can happen if a video appears twice (or more) in a feed,
e.g. on youtube where a video is added both when it appears as a
going-live-announcement and when it actually goes live.
Here we accidentally did not thread the error through the outer
`ignore_constraint_errors` before.